restructure service files
