+++ frontend/webapp/containers/setup/destination/destination.section.tsx
@@ -1,7 +1,6 @@
 import React, { useEffect, useState } from "react";
 import { useQuery } from "react-query";
 import { QUERIES, SETUP } from "@/utils/constants";
-import { getDestinationsTypes } from "@/services/setup";
 import { MONITORING_OPTIONS } from "@/components/setup/destination/utils";
 import { DestinationList, DestinationOptionMenu } from "@/components/setup";
 import Empty from "@/assets/images/empty-list.svg";
@@ -18,6 +17,7 @@ import {
 } from "./utils";
 import { KeyvalLoader } from "@/design.system";
 import { useNotification } from "@/hooks";
+import { getDestinationsTypes } from "@/services";
 
 type DestinationSectionProps = {
   sectionData: any;
