+++ frontend/webapp/services/index.tsx
@@ -1,0 +1,13 @@
+export {
+  getDestinationsTypes,
+  getDestinations,
+  getDestination,
+  setDestination,
+} from "./destinations";
+
+export {
+  getNamespaces,
+  getApplication,
+  setNamespaces,
+  getSources,
+} from "./sources";
